Moved Driver and SelfDrivingCar into the leader_board script in accordance with the usual coding style
