--- Ruby/LeaderBoard/driver.rb
@@ -1,7 +1,0 @@
-class Driver
-  attr_reader :name, :country
-  def initialize(name, country)
-    @name = name
-    @country = country
-  end
-end
