--- Ruby/LeaderBoard/self_driving_car.rb
@@ -1,9 +1,0 @@
-require_relative './driver'
-class SelfDrivingCar < Driver
-  attr_accessor :algorithm_version
-  def initialize(algorithm_version, company)
-    super
-    @algorithm_version = algorithm_version
-    @company = company
-  end
-end
